+++ src/SyndicationType/Concrete/TwitterSyndicationType.php
@@ -48,6 +48,7 @@ class TwitterSyndicationType extends AbstractSyndicationType
             [
                 'class' => '',
                 'target' => '_blank',
+                'title' => $this->translator->trans('huh.syndication_type.types.twitter.title'),
                 'onclick' => 'window.open(this.href,\'\',\'width=500,height=260,modal=yes,left=100,top=50,location=no,menubar=no,resizable=yes,scrollbars=yes,status=no,toolbar=no\');return false',
             ],
             $this