+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.1] - 2022-01-05
+- Changed: added missing title attribute to linkin, xing and twitter syndication type
+
 ## [0.5.0] - 2021-12-15
 - Added: [BREAKING!] SyndicationLinkRendererContext parameter to SyndicationLinkRenderer::renderProvider()
 - Added: [BREAKING!] SyndicationLink type property